Address review feedback on affinity group resources
Split CRUD out of resource.go into create, read, update, delete and
import files in both affinity group packages, matching the layout used
elsewhere in the provider. The move is behaviour-neutral: the unit tests
and the full acceptance suite pass unchanged.

Deprecate the cluster affinity group description attribute rather than
removing it. It is not backed by the API and has no effect, but removing
a shipped attribute breaks existing configurations, so it now carries a
deprecation message and will go in a future release.

Type server ids as int64 rather than narrowing them to int32. Compute
server ids are 64-bit, and the conversion was both unnecessary and
flagged as an overflow risk. The SDK change comes from the companion
spec fix.

Retarget the compute server managed override, which pointed at an
attribute that does not exist and so silently did nothing, and add the
missing copyright headers.
