+++ internal/sdk/oapigen/model_update_cluster_affinity_group_request_affinity_group.go
@@ -25,7 +25,7 @@ type UpdateClusterAffinityGroupRequestAffinityGroup struct {
 	// Active
 	Active *bool `json:"active,omitempty"`
 	// List of Server IDs to include in the Affinity Group
-	Servers []int32 `json:"servers,omitempty"`
+	Servers []int64 `json:"servers,omitempty"`
 	// Visibility - Set to public to allow all tenants
 	Visibility *string `json:"visibility,omitempty"`
 	// Array of tenant account ids that are allowed access
