+++ morpheus/framework/resources/cloudaffinitygroup/import.go
@@ -1,0 +1,51 @@
+// (C) Copyright 2026 Hewlett Packard Enterprise Development LP
+
+package cloudaffinitygroup
+
+import (
+	"context"
+	"fmt"
+	"strconv"
+	"strings"
+
+	"github.com/hashicorp/terraform-plugin-framework/path"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+)
+
+func (r *cloudAffinityGroupResource) ImportState(
+	ctx context.Context,
+	req resource.ImportStateRequest,
+	resp *resource.ImportStateResponse,
+) {
+	// CRITICAL BEHAVIOUR 6: Import uses composite ID cloud_id.affinity_group_id.
+	parts := strings.Split(req.ID, ".")
+	if len(parts) != 2 {
+		resp.Diagnostics.AddError("Invalid Import ID",
+			fmt.Sprintf("Expected format: cloud_id.affinity_group_id, got: %s", req.ID))
+
+		return
+	}
+
+	cloudID, err := strconv.ParseInt(parts[0], 10, 64)
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Invalid ID",
+			fmt.Sprintf("Could not parse cloud_id %q: %s", parts[0], err),
+		)
+
+		return
+	}
+
+	id, err := strconv.ParseInt(parts[1], 10, 64)
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Invalid ID",
+			fmt.Sprintf("Could not parse affinity_group_id %q: %s", parts[1], err),
+		)
+
+		return
+	}
+
+	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("cloud_id"), cloudID)...)
+	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("id"), id)...)
+}
