+++ morpheus/framework/resources/cloudaffinitygroup/resource_test.go
@@ -1,3 +1,5 @@
+// (C) Copyright 2026 Hewlett Packard Enterprise Development LP
+
 package cloudaffinitygroup_test
 
 import (
