+++ morpheus/framework/resources/clusteraffinitygroup/delete.go
@@ -1,0 +1,57 @@
+// (C) Copyright 2026 Hewlett Packard Enterprise Development LP
+
+package clusteraffinitygroup
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+
+	"github.com/HPE/terraform-provider-hpe/morpheus/utils/constants"
+	"github.com/HPE/terraform-provider-hpe/morpheus/utils/errfmt"
+	"github.com/HPE/terraform-provider-hpe/morpheus/utils/versioncheck"
+)
+
+func (r *clusterAffinityGroupResource) Delete(
+	ctx context.Context,
+	req resource.DeleteRequest,
+	resp *resource.DeleteResponse,
+) {
+	client, err := r.NewClient(ctx)
+	if err != nil {
+		errfmt.DiagClientError(&resp.Diagnostics, err)
+
+		return
+	}
+
+	// MORPH-15506 appliance version gate — see Create. Delete is gated too, so
+	// the contract is simply "this provider does not manage affinity groups
+	// below 8.0.10" rather than a half-state where destroy is the one operation
+	// that still works. Read is gated anyway, so a refreshing destroy already
+	// stops here; a practitioner holding an orphaned group on an older
+	// appliance removes it with `terraform state rm` and the Morpheus UI.
+	resp.Diagnostics.Append(versioncheck.Require(
+		ctx, client, gatedFeature, constants.AffinityGroupMinVersion,
+	)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	var state ClusterAffinityGroupModel
+	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	clusterID := state.ClusterId.ValueInt64()
+	id := state.Id.ValueInt64()
+
+	_, httpResp, err := client.ClustersAPI.DeleteClusterAffinityGroup(ctx, clusterID, id).Execute()
+	if err := errfmt.CheckResponse(err, httpResp); err != nil {
+		errfmt.DiagError(
+			&resp.Diagnostics, errfmt.OpDelete, "cluster_affinity_group", "", err, httpResp,
+		)
+
+		return
+	}
+}
