+++ morpheus/framework/resources/clusteraffinitygroup/import.go
@@ -1,0 +1,50 @@
+// (C) Copyright 2026 Hewlett Packard Enterprise Development LP
+
+package clusteraffinitygroup
+
+import (
+	"context"
+	"fmt"
+	"strconv"
+	"strings"
+
+	"github.com/hashicorp/terraform-plugin-framework/path"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+)
+
+func (r *clusterAffinityGroupResource) ImportState(
+	ctx context.Context,
+	req resource.ImportStateRequest,
+	resp *resource.ImportStateResponse,
+) {
+	parts := strings.Split(req.ID, ".")
+	if len(parts) != 2 {
+		resp.Diagnostics.AddError("Invalid Import ID",
+			fmt.Sprintf("Expected format: cluster_id.affinity_group_id, got: %s", req.ID))
+
+		return
+	}
+
+	clusterID, err := strconv.ParseInt(parts[0], 10, 64)
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Invalid ID",
+			fmt.Sprintf("Could not parse cluster_id %q: %s", parts[0], err),
+		)
+
+		return
+	}
+
+	id, err := strconv.ParseInt(parts[1], 10, 64)
+	if err != nil {
+		resp.Diagnostics.AddError(
+			"Invalid ID",
+			fmt.Sprintf("Could not parse affinity_group_id %q: %s", parts[1], err),
+		)
+
+		return
+	}
+
+	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("cluster_id"), clusterID)...)
+	resp.Diagnostics.Append(resp.State.SetAttribute(ctx, path.Root("id"), id)...)
+}
