+++ morpheus/framework/resources/storagebucket/resource.go
@@ -1,3 +1,5 @@
+// (C) Copyright 2026 Hewlett Packard Enterprise Development LP
+
 package storagebucket
 
 import (
